separating geometry classes
