+++ src/main/scala/net/psforever/objects/ExplosiveDeployable.scala
@@ -6,7 +6,7 @@ import net.psforever.objects.ce._
 import net.psforever.objects.definition.DeployableDefinition
 import net.psforever.objects.definition.converter.SmallDeployableConverter
 import net.psforever.objects.equipment.JammableUnit
-import net.psforever.objects.geometry.Geometry3D
+import net.psforever.objects.geometry.d3.Geometry3D
 import net.psforever.objects.serverobject.affinity.FactionAffinity
 import net.psforever.objects.serverobject.PlanetSideServerObject
 import net.psforever.objects.serverobject.damage.{Damageable, DamageableEntity}
